Fail early if redis is unreachable and move function to its own module
